+++ spec/api/endpoints/slack_endpoint_spec.rb
@@ -396,6 +396,8 @@ describe Api::Endpoints::SlackEndpoint do
                event_time: 1_547_842_101,
                authed_users: ['U04KB5WQR']
           expect(last_response.status).to eq 201
+
+          expect(activity.reload.bragged_at).to_not be nil
         end
       end
     end